Refactor main into app modules and route registrars
Split large main.py into focused application modules and move app setup logic out of the monolith. Added an app factory, service container, consolidated lifespan (startup/shutdown) logic, ASGI request-logging middleware, and route registrars (internal, public_v1, MCP mount). Introduced health endpoints (liveness + OpenSearch readiness), service initialization wiring, startup orchestrator and helper utilities (default docs service, JWT keygen, OpenSearch init, URL content fetcher). main.py was slimmed to re-export essentials and delegate app creation to app.factory.create_app. This improves modularity, testability, and separates concerns for lifecycle, routing, and service construction.
